// packages/codegen-kernel/src/pluginTypes.ts
export type Framework = "Flutter" | "SwiftUI" | "HTML" | "Tailwind" | "Compose";
export interface HTMLSettings {
  showLayerNames: boolean;
  embedImages: boolean;
  embedVectors: boolean;
  useColorVariables: boolean;
  htmlGenerationMode: "html" | "jsx" | "styled-components" | "svelte";
}
export interface TailwindSettings extends HTMLSettings {
  tailwindGenerationMode: "html" | "jsx" | "twig";
  roundTailwindValues: boolean;
  roundTailwindColors: boolean;
  useColorVariables: boolean;
  customTailwindPrefix?: string;
  embedVectors: boolean;
  baseFontSize: number;
  useTailwind4: boolean;
  thresholdPercent: number;
  baseFontFamily: string;
  fontFamilyCustomConfig: Record<string, string[]>
}
export interface FlutterSettings {
  flutterGenerationMode: "fullApp" | "stateless" | "snippet";
}
export interface SwiftUISettings {
  swiftUIGenerationMode: "preview" | "struct" | "snippet";
}
export interface ComposeSettings {
  composeGenerationMode: "snippet" | "composable" | "screen";
}
export interface PluginSettings
  extends HTMLSettings,
    TailwindSettings,
    FlutterSettings,
    SwiftUISettings,
    ComposeSettings {
  framework: Framework;
  useOldPluginVersion2025: boolean;
  responsiveRoot: boolean;
}